+++ docs/source/user_guide/changelog.rst
@@ -9,6 +9,7 @@ Changelog
 - Trajectories in csv format contain the time in seconds in the first column.
 - Remove CONVERT_TRAJECTORIES_TO_CSV_AND_JSON, set to always True.
 - Reorganize trajectories output folder.
+- Change video extension limitation for everything OpenCV can read.
 
 5.1.6
 =====